ce-p2ios-xcframework: add build_engine.sh + fix UniFFI 0.30 compat
build_engine.sh:
- Builds staticlib for ios-arm64, ios-sim (arm64+x86_64 fat), macos-arm64
- Generates UniFFI Swift bindings via macOS debug dylib
- Packages ConstructEngine.xcframework via xcodebuild -create-xcframework
- Copies construct_engine.swift to ConstructMessenger/
- Usage: ./build_engine.sh [--ios|--sim|--mac|--bindings|--clean|--debug]

UniFFI 0.30 fixes:
- Constructor returns Self (not Arc<Self>) — 0.30 wraps in Arc internally
- event_buffer: usize → u64 to match UDL u64 declaration
- Add auth_token: Option<String> to UDL EngineConfig dictionary
- Add Unauthenticated(string) to UDL EngineError enum

+++ src/config.rs
@@ -38,7 +38,7 @@ pub struct EngineConfig {
     pub masque_port: Option<u16>,
 
     /// Event channel buffer depth (default: 1024)
-    pub event_buffer: usize,
+    pub event_buffer: u64,
 }
 
 impl Default for EngineConfig {
